requirements.txt
# Core API dependencies
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic>=2.9.2
python-multipart>=0.0.20

# Audio and ML dependencies
torch>=2.0.0
torchaudio>=2.0.0
torchcodec>=0.0.1
numpy>=1.24.0
librosa>=0.10.0
soundfile>=0.12.0

# TTS dependencies
coqui-tts>=0.21.0

# GLM-TTS dependencies (needed for glm-tts backend)
onnxruntime-gpu>=1.17.0
HyperPyYAML>=1.2.0
safetensors>=0.4.0
peft>=0.7.0
inflect>=7.0.0
contractions>=0.1.70
WeTextProcessing>=1.0.3
pypinyin>=0.44.0
jieba>=0.42.1
emoji>=2.0.0
pronouncing>=0.2.0
sentencepiece>=0.1.99
zhconv>=1.4.3
zhon>=2.1.1
pydub>=0.25.1
x_transformers>=1.40.0
einops>=0.8.0
jiwer>=4.0.0

# Speech dependencies for OpenAudio TTS
transformers>=4.30.0
tokenizers>=0.13.0
accelerate>=0.20.0
datasets>=2.0.0

# Additional audio processing
scipy>=1.10.0

# HTTP client for requests
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.8.0

# YAML for configuration
PyYAML>=6.0

# Logging and utilities
colorlog>=6.7.0

# WebSocket support for real-time TTS
websockets>=11.0.0
ormsgpack>=1.4.0

# STT for testing TTS output
SpeechRecognition>=3.10.0
openai-whisper>=20231117
pyaudio>=0.2.11

# System monitoring
psutil>=5.9.0

# Prometheus metrics
prometheus-client>=0.19.0

# Development dependencies
packaging>=21.0
setuptools>=65.0